+++ src/source_controller/objectcontroller/service/set.go
@@ -10,7 +10,7 @@
  * limitations under the License.
  */
 
-package openapi
+package service
 
 import (
 	"encoding/json"
@@ -23,64 +23,47 @@ import (
 	"gopkg.in/mgo.v2/bson"
 
 	"configcenter/src/common"
-	"configcenter/src/common/base"
 	"configcenter/src/common/blog"
-	"configcenter/src/common/core/cc/actions"
 	"configcenter/src/common/metadata"
 	"configcenter/src/common/util"
 	"configcenter/src/source_controller/common/eventdata"
 )
 
-var set *setAction = &setAction{}
-
-type setAction struct {
-	base.BaseAction
-}
-
-func init() {
-
-	actions.RegisterNewAction(actions.Action{Verb: common.HTTPDelete, Path: "/openapi/set/delhost", Params: nil, Handler: set.DeleteSetHost})
-
-	// create CC object
-	set.CreateAction()
-}
-
-func (cli *setAction) DeleteSetHost(req *restful.Request, resp *restful.Response) {
+func (cli *Service) DeleteSetHost(req *restful.Request, resp *restful.Response) {
 	// get the language
 	language := util.GetActionLanguage(req)
 	// get the error factory by the language
-	defErr := cli.CC.Error.CreateDefaultCCErrorIf(language)
+	defErr := cli.Core.CCErr.CreateDefaultCCErrorIf(language)
 
-	cli.CallResponseEx(func() (int, interface{}, error) {
+	blog.Debug("DeleteSetHost start !")
+	value, err := ioutil.ReadAll(req.Request.Body)
+	if nil != err {
+		blog.Error("read request body failed, error:%v", err)
+		resp.WriteError(http.StatusBadRequest, &meta.RespError{Msg: defErr.New(common.CCErrCommHTTPReadBodyFailed, err.Error())})
+		return
+	}
 
-		blog.Debug("DeleteSetHost start !")
-		value, err := ioutil.ReadAll(req.Request.Body)
-		if nil != err {
-			blog.Error("read request body failed, error:%v", err)
-			return http.StatusInternalServerError, nil, defErr.Error(common.CCErrCommHTTPReadBodyFailed)
-		}
+	blog.Debug("DeleteSetHost http body data: %s", value)
+	input := make(map[string]interface{})
+	err = json.Unmarshal(value, &input)
+	if nil != err {
+		blog.Error("unmarshal json error:%v", err)
+		resp.WriteError(http.StatusBadRequest, &meta.RespError{Msg: defErr.New(common.CCErrCommJSONUnmarshalFailed, err.Error())})
+		return
+	}
 
-		blog.Debug("DeleteSetHost http body data: %s", value)
-		input := make(map[string]interface{})
-		err = json.Unmarshal(value, &input)
-		if nil != err {
-			blog.Error("unmarshal json error:%v", err)
-			return http.StatusInternalServerError, nil, defErr.Error(common.CCErrCommJSONUnmarshalFailed)
-		}
+	err = delModuleConfigSet(input, req)
+	if err != nil {
+		blog.Error("fail to delSetConfigHost: %v", err)
+		resp.WriteError(http.StatusBadRequest, &meta.RespError{Msg: defErr.New(common.CCErrCommParamsInvalid, err.Error())})
+		return
+	}
 
-		err = delModuleConfigSet(input, req)
-		if err != nil {
-			blog.Error("fail to delSetConfigHost: %v", err)
-			return http.StatusInternalServerError, nil, defErr.Error(common.CCErrCommParamsInvalid)
-
-		}
-
-		return http.StatusOK, nil, nil
-	}, resp)
+	resp.WriteEntity(meta.Response{BaseResp: meta.SuccessBaseResp})
 }
 
 // TODO
-func getModuleConfigCount(con map[string]interface{}) (int, error) {
+func (cli *Service) getModuleConfigCount(con map[string]interface{}) (int, error) {
 	count, err := set.CC.InstCli.GetCntByCondition("cc_ModuleHostConfig", con)
 	if err != nil {
 		blog.Error("fail getModuleConfigCount error:%v", err)
@@ -89,7 +72,7 @@ func getModuleConfigCount(con map[string]interface{}) (int, error) {
 	return count, err
 }
 
-func delModuleConfigSet(input map[string]interface{}, req *restful.Request) error {
+func (cli *Service) delModuleConfigSet(input map[string]interface{}, req *restful.Request) error {
 	tableName := "cc_ModuleHostConfig"
 
 	appID, ok := input[common.BKAppIDField]
@@ -115,19 +98,7 @@ func delModuleConfigSet(input map[string]interface{}, req *restful.Request) erro
 		blog.Error("fail to delSetConfigHost: %v", err)
 		return err
 	}
-	/*count, err := getModuleConfigCount(input)
-	if err != nil {
-		blog.Error("fail to delSetConfigHost: %v", err)
-		return err
-	}
-	blog.Debug("count:%v", count)
-	if count != 0 {
-		err = delModuleConfigSet(input, req)
-		if err != nil {
-			blog.Error("fail to delSetConfigHost: %v", err)
-			return err
-		}
-	}*/
+
 	//发送删除主机关系事件
 	ec := eventdata.NewEventContextByReq(req)
 	for oldContent := range oldContents {
@@ -192,7 +163,7 @@ func delModuleConfigSet(input map[string]interface{}, req *restful.Request) erro
 	return nil
 }
 
-func GetIdleModule(appID interface{}) (interface{}, interface{}, error) {
+func (cli *Service) GetIdleModule(appID interface{}) (interface{}, interface{}, error) {
 	params := common.KvMap{common.BKAppIDField: appID, common.BKDefaultField: common.DefaultResModuleFlag, common.BKModuleNameField: common.DefaultResModuleName}
 	var result bson.M
 	err := set.CC.InstCli.GetOneByCondition("cc_ModuleBase", []string{common.BKModuleIDField, common.BKSetIDField}, params, &result)
